+++ packages/build-tools/create-webpack-config.js
@@ -1,37 +1,27 @@
-const path = require('path');
-const webpack = require('webpack');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const HardSourceWebpackPlugin = require('hard-source-webpack-plugin-patch');
-const TerserPlugin = require('terser-webpack-plugin');
-const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
-const autoprefixer = require('autoprefixer');
-const postcssDiscardDuplicates = require('postcss-discard-duplicates');
-const ManifestPlugin = require('webpack-manifest-plugin');
 const fs = require('fs');
+const path = require('path');
 const deepmerge = require('deepmerge');
-const resolve = require('resolve');
+const webpack = require('webpack');
+const { merge } = require('webpack-merge');
+
+// Plugins/loaders
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const CssMinimizerPlugin = require('css-minimizer-webpack-plugin');
+const { WebpackManifestPlugin } = require('webpack-manifest-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
-const npmSass = require('npm-sass');
-const merge = require('webpack-merge');
-const SassDocPlugin = require('@bolt/sassdoc-webpack-plugin');
+const SpriteLoaderPlugin = require('svg-sprite-loader/plugin');
+const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
+const postcssDiscardDuplicates = require('postcss-discard-duplicates');
+const autoprefixer = require('autoprefixer');
+
+// Helpers/config
 const { getConfig } = require('@bolt/build-utils/config-store');
 const { boltWebpackProgress } = require('@bolt/build-utils/webpack-helpers');
-const SpriteLoaderPlugin = require('svg-sprite-loader/plugin');
-const {
-  webpackStats,
-  statsPreset,
-} = require('@bolt/build-utils/webpack-verbosity');
-
+const { getBoltManifest } = require('@bolt/build-utils/manifest');
+const npmSass = require('npm-sass'); // @todo: Remove when we switch to Dart sass
+const sassExportData = require('@bolt/sass-export-data'); // @todo: Solve this problem without node-sass
 const babelConfig = require('@bolt/babel-preset-bolt');
-
-const {
-  getBoltManifest,
-  mapComponentNameToTwigNamespace,
-} = require('@bolt/build-utils/manifest');
 const log = require('@bolt/build-utils/log');
-
-// Store set of webpack configs used in multiple builds
-let webpackConfigs = [];
 
 async function createWebpackConfig(buildConfig) {
   const config = buildConfig;
@@ -44,16 +34,6 @@ async function createWebpackConfig(buildConfig) {
     : config.wwwDir
     ? `/${path.relative(config.wwwDir, config.buildDir)}/`
     : config.buildDir; // @todo Ensure ends with `/` or we can get `distfonts/` instead of `dist/fonts/`
-
-  // @TODO: move this setting to .boltrc config
-  const sassExportData = require('@bolt/sass-export-data')({
-    path: config.dataDir,
-  });
-
-  // map out Twig namespaces with the NPM package name
-
-  // filename suffix to tack on based on lang being compiled for
-  let langSuffix = `${config.lang ? '-' + config.lang : ''}`;
 
   /**
    * Build WebPack config's `entry` object
@@ -98,11 +78,7 @@ async function createWebpackConfig(buildConfig) {
 
   function getSassLoaders() {
     // Default global Sass data defined
-    let globalSassData = [
-      `$bolt-namespace: ${config.namespace};`,
-      // output $bolt-lang variable in Sass even if not specified so things fall back accordingly.
-      `${config.lang ? `$bolt-lang: ${config.lang};` : '$bolt-lang: null;'}`,
-    ];
+    let globalSassData = [`$bolt-namespace: ${config.namespace};`];
 
     // Merge together global Sass data overrides specified in a .boltrc config
     if (config.globalData.scss && config.globalData.scss.length !== 0) {
@@ -127,34 +103,30 @@ async function createWebpackConfig(buildConfig) {
         loader: 'css-loader',
         options: {
           sourceMap: config.sourceMaps,
-          modules: false, // needed for JS referencing classNames directly, such as critical fonts
+          // needed for JS referencing classNames directly, such as critical fonts
+          // @todo: see if we still need this now that critical css has been removed
+          modules: false,
         },
       },
       {
         loader: 'postcss-loader',
         options: {
           sourceMap: config.sourceMaps,
-          plugins: () => [
-            postcssDiscardDuplicates,
-            autoprefixer({
-              grid: true,
-            }),
-          ],
-        },
-      },
-      {
-        loader: 'clean-css-loader',
-        options: {
-          level: config.prod ? 1 : 0,
-          format: config.prod ? false : 'beautify',
-          inline: ['remote'],
+          postcssOptions: {
+            plugins: [
+              postcssDiscardDuplicates,
+              // @todo: Consider switching to postcss-preset-env which polyfills modern CSS
+              autoprefixer,
+            ],
+          },
         },
       },
       {
         loader: 'resolve-url-loader',
       },
-
       {
+        // Note: intentionally not upgrading sass-loader yet. Compile time was
+        // ~1.5x slower after trial upgrade. Wait until we switch to Dart.
         loader: 'sass-loader',
         options: {
           sourceMap: config.sourceMaps,
@@ -162,7 +134,9 @@ async function createWebpackConfig(buildConfig) {
           sassOptions: {
             outputStyle: 'nested',
             importer: [npmSass.importer],
-            functions: sassExportData,
+            functions: sassExportData({
+              path: config.dataDir,
+            }),
             precision: 3,
           },
         },
@@ -184,11 +158,6 @@ async function createWebpackConfig(buildConfig) {
         '.tsx',
         '.jpg',
       ],
-      alias: {
-        react: 'preact/compat',
-        'react-dom/test-utils': 'preact/test-utils',
-        'react-dom': 'preact/compat',
-      },
     },
     module: {
       rules: [
@@ -206,28 +175,19 @@ async function createWebpackConfig(buildConfig) {
         },
         {
           test: /\.(woff|woff2)$/,
-          use: [
-            {
-              loader: 'url-loader',
-              options: {
-                limit: 500,
-                name: 'fonts/[name].[ext]',
-              },
-            },
-          ],
+          type: 'asset/resource',
+          generator: {
+            filename: 'fonts/[name][ext]',
+          },
         },
         {
           test: /\.svg$/,
           oneOf: [
             {
               issuer: /\.scss$/,
+              // @see: https://dev.to/smelukov/webpack-5-asset-modules-2o3h
+              type: 'asset/resource',
               use: [
-                {
-                  loader: 'file-loader',
-                  options: {
-                    name: '[name].[ext]',
-                  },
-                },
                 {
                   loader: 'svgo-loader',
                   options: {
@@ -240,10 +200,7 @@ async function createWebpackConfig(buildConfig) {
               use: [
                 {
                   loader: 'babel-loader',
-                  options: {
-                    babelrc: false,
-                    presets: [babelConfig],
-                  },
+                  options: babelConfig,
                 },
                 {
                   loader: 'svg-sprite-loader',
@@ -271,14 +228,7 @@ async function createWebpackConfig(buildConfig) {
         },
         {
           test: /\.(cur|png|jpg)$/,
-          use: [
-            {
-              loader: 'file-loader',
-              options: {
-                name: '[name].[ext]',
-              },
-            },
-          ],
+          type: 'asset/resource',
         },
         {
           test: [/\.yml$/, /\.yaml$/],
@@ -286,27 +236,15 @@ async function createWebpackConfig(buildConfig) {
         },
         {
           test: [/\.html$/],
-          loader: 'raw-loader', // file as string
+          type: 'asset/source',
         },
       ],
     },
     mode: config.prod ? 'production' : 'development',
+    cache: config.enableCache,
     optimization: {
       sideEffects: true,
       usedExports: true,
-      minimizer: config.prod
-        ? [
-            new TerserPlugin({
-              test: /\.m?js(\?.*)?$/i,
-              sourceMap: config.sourceMaps,
-              cache: true,
-              parallel: true,
-              terserOptions: {
-                safari10: true,
-              },
-            }),
-          ]
-        : [],
     },
     plugins: [
       new SpriteLoaderPlugin({
@@ -317,22 +255,17 @@ async function createWebpackConfig(buildConfig) {
         },
       }),
       new webpack.ProgressPlugin(boltWebpackProgress), // Ties together the Bolt custom Webpack messages + % complete
-      new webpack.NoEmitOnErrorsPlugin(),
     ],
   };
 
   if (config.prod) {
-    // https://webpack.js.org/plugins/module-concatenation-plugin/
-    sharedWebpackConfig.plugins.push(
-      new webpack.optimize.ModuleConcatenationPlugin(),
-    );
-
-    // Optimize CSS - https://github.com/NMFR/optimize-css-assets-webpack-plugin
-    sharedWebpackConfig.plugins.push(
-      new OptimizeCssAssetsPlugin({
-        canPrint: config.verbosity > 2,
-        cssProcessor: require('cssnano'),
-        cssProcessorPluginOptions: {
+    sharedWebpackConfig.devtool =
+      config.sourceMaps === false ? false : 'hidden-source-map';
+    sharedWebpackConfig.optimization.minimize = true;
+    sharedWebpackConfig.optimization.minimizer = [
+      `...`,
+      new CssMinimizerPlugin({
+        minimizerOptions: {
           preset: [
             'default',
             {
@@ -346,16 +279,12 @@ async function createWebpackConfig(buildConfig) {
           ],
         },
       }),
-    );
-
-    // @todo evaluate best source map approach for production builds -- particularly source-map vs hidden-source-map
-    sharedWebpackConfig.devtool =
-      config.sourceMaps === false ? '' : 'hidden-source-map';
+    ];
   } else {
     // not prod
     // @todo fix source maps
     sharedWebpackConfig.devtool =
-      config.sourceMaps === false ? '' : 'eval-source-map';
+      config.sourceMaps === false ? false : 'eval-source-map';
   }
 
   // Simple Configuration
@@ -375,6 +304,10 @@ async function createWebpackConfig(buildConfig) {
     sharedWebpackConfig = merge(sharedWebpackConfig, config.configureWebpack);
   }
 
+  if (config.analyze) {
+    sharedWebpackConfig.plugins.push(new BundleAnalyzerPlugin());
+  }
+
   // Generate global JS data based on if the build is for ES Module-supporting browsers or not
   function getGlobalJSData() {
     let globalJsData = {
@@ -389,7 +322,6 @@ async function createWebpackConfig(buildConfig) {
         namespace: JSON.stringify(config.namespace),
         config: {
           prod: config.prod,
-          lang: JSON.stringify(config.lang),
           env: JSON.stringify(config.env),
         },
       },
@@ -413,31 +345,36 @@ async function createWebpackConfig(buildConfig) {
     return globalJsData;
   }
 
-  const webpackConfig = merge(sharedWebpackConfig, {
-    entry: await buildWebpackEntry(true),
+  const mergedConfig = merge(sharedWebpackConfig, {
+    entry: await buildWebpackEntry(),
     resolve: {
       mainFields: ['esnext', 'jsnext:main', 'browser', 'module', 'main'],
     },
     output: {
-      futureEmitAssets: true,
       path: path.resolve(process.cwd(), config.buildDir),
       // @todo: switch this to output .client.js and .server.js file prefixes when we hit Bolt v3.0
-      filename: `[name]${langSuffix}${
-        config.mode !== 'client' ? `.${config.mode}` : ''
-      }.js`,
-      chunkFilename: `[name]-bundle${langSuffix}-[chunkhash].js`,
+      filename: `[name]${config.mode !== 'client' ? `.${config.mode}` : ''}.js`,
+      chunkFilename: `[name]-bundle-[chunkhash].js`,
       publicPath,
+    },
+    // Emulate Drupal environment, where react/react-dom is on the window (temporary until DS-863 is done)
+    externals: {
+      react: 'React',
+      'react-dom': 'ReactDOM',
     },
     plugins: [
       new webpack.DefinePlugin(getGlobalJSData(true)),
-      new CopyWebpackPlugin(config.copy ? config.copy : []),
+      // CopyWebpackPlugin throws an error if you don't pass it a configuration object
+      config.copy
+        ? new CopyWebpackPlugin({ patterns: config.copy })
+        : undefined,
       new MiniCssExtractPlugin({
-        filename: `[name]${langSuffix}.css`,
-        chunkFilename: `[id]${langSuffix}.css`,
+        filename: `[name].css`,
+        chunkFilename: `[id].css`,
       }),
       // @todo This needs to be in `config.dataDir`
-      new ManifestPlugin({
-        fileName: `bolt-webpack-manifest${langSuffix}${
+      new WebpackManifestPlugin({
+        fileName: `bolt-webpack-manifest${
           config.mode === 'client' ? '' : `.${config.mode}`
         }.json`,
         publicPath,
@@ -446,7 +383,7 @@ async function createWebpackConfig(buildConfig) {
           name: 'Bolt Modern Manifest',
         },
       }),
-    ],
+    ].filter(item => item !== undefined),
     module: {
       rules: [
         {
@@ -457,10 +394,7 @@ async function createWebpackConfig(buildConfig) {
           use: [
             {
               loader: 'babel-loader',
-              options: {
-                babelrc: false,
-                presets: [babelConfig],
-              },
+              options: babelConfig,
             },
           ],
         },
@@ -476,11 +410,10 @@ async function createWebpackConfig(buildConfig) {
             },
             {
               // no issuer here as it has a bug when its an entry point - https://github.com/webpack/webpack/issues/5906
-              use: [
-                // 'css-hot-loader',
-                MiniCssExtractPlugin.loader,
-                getSassLoaders(true),
-              ].reduce((acc, val) => acc.concat(val), []),
+              use: [MiniCssExtractPlugin.loader, getSassLoaders(true)].reduce(
+                (acc, val) => acc.concat(val),
+                [],
+              ),
             },
           ],
         },
@@ -488,72 +421,40 @@ async function createWebpackConfig(buildConfig) {
     },
   });
 
-  // cache mode significantly speeds up subsequent build times
-  if (config.enableCache) {
-    webpackConfig.plugins.push(
-      new HardSourceWebpackPlugin({
-        info: {
-          level: 'warn',
-        },
-        cacheDirectory: path.join(process.cwd(), `./cache/webpack`),
-        // Clean up large, old caches automatically.
-        cachePrune: {
-          // Caches younger than `maxAge` are not considered for deletion. They must
-          // be at least this (default: 2 days) old in milliseconds.
-          maxAge: 2 * 24 * 60 * 60 * 1000,
-          // All caches together must be larger than `sizeThreshold` before any
-          // caches will be deleted. Together they must be at least 300MB in size
-          sizeThreshold: 300 * 1024 * 1024,
-        },
-      }),
-    );
+  if (config.env === 'drupal') {
+    const themeRegExp = new RegExp('^(' + config.themeNames.join('|') + ')-');
+    const disablePlugins = ['ManifestPlugin'];
+
+    // Remove any empty entries (for example one that just has a Twig template)
+    for (const i in mergedConfig.entry) {
+      if (!mergedConfig.entry[i].length) {
+        delete mergedConfig.entry[i];
+      }
+    }
+
+    const entries = {};
+    for (const name in mergedConfig.entry) {
+      // Replace "-" with "/", e.g. `@pega_bolt_theme-components-wysiwyg` => `@pega_bolt_theme/components-wysiwyg`
+      // Replace "@" with "", e.g. `@pega_bolt_theme/components-wysiwyg` => `pega_bolt_theme/components-wysiwyg`
+      const updatedName = name.replace(themeRegExp, '$1/').replace(/@/g, '');
+      entries[updatedName] = mergedConfig.entry[name];
+    }
+    mergedConfig.entry = entries;
+
+    mergedConfig.plugins = mergedConfig.plugins.filter(plugin => {
+      return !disablePlugins.includes(plugin.constructor.name);
+    });
   }
 
-  let outputConfig = [];
-
-  outputConfig.push(webpackConfig);
-
-  return outputConfig;
-}
-
-// Helper function to associate each unique language in the build config with a separate Webpack build instance (making filenames, etc unique);
-async function assignLangToWebpackConfig(config, lang) {
-  let langSpecificConfig = config;
-
-  if (lang) {
-    langSpecificConfig.lang = lang; // Make sure only ONE language config is set per Webpack build instance.
-  }
-
-  let langSpecificWebpackConfigs = await createWebpackConfig(
-    langSpecificConfig,
-  );
-
-  langSpecificWebpackConfigs.forEach(langSpecificWebpackConfig => {
-    webpackConfigs.push(langSpecificWebpackConfig);
-  });
+  return mergedConfig;
 }
 
 module.exports = async function() {
-  const config = await getConfig();
+  const boltConfig = await getConfig();
+  const webpackConfig = await createWebpackConfig(boltConfig);
 
-  return new Promise(async (resolve, reject) => {
-    const langs = config.lang;
-    const promises = [];
-
-    // update the array of Webpack configs so each config is assigned to only one language (used in the filename's suffix when bundling language-tailed CSS and JS)
-    if (Array.isArray(langs)) {
-      for (const lang of langs) {
-        /* eslint-disable no-await-in-loop */
-        promises.push(await assignLangToWebpackConfig(config, lang));
-      }
-    } else if (langs === 'en') {
-      promises.push(await assignLangToWebpackConfig(config, null));
-    } else {
-      promises.push(await assignLangToWebpackConfig(config, config.lang));
-    }
-
-    await Promise.all(promises).then(() => {
-      return resolve(webpackConfigs);
-    });
-  });
+  // Bolt still expects an array of configs even though we only pass one,
+  // a hold over from the multi-lang support.
+  // @todo: Update this to handle just a single configuration object.
+  return [webpackConfig];
 };