+++ packages/build-tools/tasks/webpack-tasks.js
@@ -2,7 +2,6 @@ const webpack = require('webpack');
 const express = require('express');
 const browserSync = require('browser-sync').create();
 const webpackDevMiddleware = require('webpack-dev-middleware');
-const webpackHotMiddleware = require('webpack-hot-middleware');
 const chalk = require('chalk');
 const opn = require('better-opn');
 const { handleRequest } = require('@bolt/api');
@@ -35,6 +34,9 @@ async function compile(customWebpackConfig) {
   return new Promise((resolve, reject) => {
     const compiler = boltWebpackMessages(webpack(webpackConfig));
     compiler.run((err, stats) => {
+      // Don't forget to close the compiler
+      // @see https://webpack.js.org/api/node/#run
+      compiler.close();
       if (err) {
         return reject(err);
       } else {
@@ -55,10 +57,23 @@ async function watch(customConfig) {
   return new Promise((resolve, reject) => {
     const compiler = boltWebpackMessages(webpack(webpackConfig));
 
-    compiler.watch({
-      // https://webpack.js.org/configuration/watch/#watchoptions
-      aggregateTimeout: 300,
-    });
+    compiler.watch(
+      {
+        // https://webpack.js.org/configuration/watch/#watchoptions
+        aggregateTimeout: 300,
+        poll: 1000,
+      },
+      (err, stats) => {
+        if (err) {
+          console.error(err);
+          return;
+        }
+
+        console.log(
+          stats.toString(statsPreset(webpackStats[boltBuildConfig.verbosity])),
+        );
+      },
+    );
   });
 }
 
@@ -144,10 +159,7 @@ async function server(customWebpackConfig) {
     );
     app.use(
       webpackDevMiddleware(compiler, {
-        quiet: true,
-        stats: 'errors-warnings',
         writeToDisk: true,
-        logLevel: 'error',
         stats: statsPreset(webpackStats[boltBuildConfig.verbosity]),
       }),
     );