+++ packages/configs/babel-preset-bolt/index.js
@@ -3,36 +3,12 @@ module.exports = {
     [
       '@babel/preset-env',
       {
-        ignoreBrowserslistConfig: true,
-        modules: false,
-        debug: false,
-        corejs: 3,
-        useBuiltIns: 'entry',
         targets: {
           browsers: require('@bolt/browserslist-config'),
         },
       },
     ],
+    '@babel/preset-react',
   ],
-  plugins: [
-    '@babel/plugin-proposal-optional-chaining',
-    [
-      '@babel/plugin-proposal-decorators',
-      {
-        decoratorsBeforeExport: true,
-      },
-    ],
-    ['@babel/plugin-proposal-class-properties', { loose: true }],
-    '@babel/plugin-syntax-jsx',
-    [
-      '@babel/plugin-transform-react-jsx',
-      {
-        pragma: 'h',
-        pragmaFrag: 'Fragment',
-        throwIfNamespace: false,
-        useBuiltIns: false,
-      },
-    ],
-    '@babel/plugin-proposal-nullish-coalescing-operator',
-  ],
+  plugins: [['@babel/plugin-proposal-decorators', { version: 'legacy' }]],
 };