+++ packages/core-v3.x/data/index.js
@@ -1,2 +1,1 @@
-export * from './image-sizes';
 export * from './spacing-sizes';