+++ packages/core-v3.x/renderers/index.js
@@ -1,2 +1,2 @@
-export * from './renderer-preact';
+export * from './renderer-react';
 export * from './renderer-lit-html';