+++ packages/website-ui/uikit-workshop/src/scripts/actions/app.js
@@ -1,6 +1,7 @@
 export const UPDATE_THEME_MODE = 'UPDATE_THEME_MODE';
 export const UPDATE_TEST_MODE = 'UPDATE_TEST_MODE';
-export const UPDATE_ARCHIVE_MODE = 'UPDATE_ARCHIVE_MODE';
+export const UPDATE_ARCHIVE_MODE = 'UPDATE_DRUPAL_MODE';
+export const UPDATE_DRUPAL_MODE = 'UPDATE_ARCHIVE_MODE';
 export const UPDATE_LAYOUT_MODE = 'UPDATE_LAYOUT_MODE';
 export const UPDATE_DRAWER_ANIMATION_STATE = 'UPDATE_DRAWER_ANIMATION_STATE';
 export const UPDATE_DRAWER_STATE = 'UPDATE_DRAWER_STATE';
@@ -79,6 +80,15 @@ export const updateArchiveMode = archiveMode => (dispatch, getState) => {
     dispatch({
       type: UPDATE_ARCHIVE_MODE,
       archiveMode,
+    });
+  }
+};
+
+export const updateDrupalMode = drupalMode => (dispatch, getState) => {
+  if (getState().app.drupalMode !== drupalMode) {
+    dispatch({
+      type: UPDATE_DRUPAL_MODE,
+      drupalMode,
     });
   }
 };