+++ packages/components/bolt-listing-teaser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.9.0](https://github.com/boltdesignsystem/bolt/tree/master/packages/components/bolt-listing-teaser/compare/v5.8.2...v5.9.0) (2022-12-02)
+
+**Note:** Version bump only for package @bolt/components-listing-teaser
+
+
+
+
+
+# [5.8.0](https://github.com/boltdesignsystem/bolt/tree/master/packages/components/bolt-listing-teaser/compare/v5.7.5...v5.8.0) (2022-10-18)
+
+**Note:** Version bump only for package @bolt/components-listing-teaser
+
+
+
+
+
 ## [5.7.4](https://github.com/boltdesignsystem/bolt/tree/master/packages/components/bolt-listing-teaser/compare/v5.7.3...v5.7.4) (2022-09-27)
 
 **Note:** Version bump only for package @bolt/components-listing-teaser