+++ packages/website-ui/uikit-workshop/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.9.0](https://github.com/bolt-design-system/bolt/tree/master/packages/uikit-workshop/compare/v5.8.2...v5.9.0) (2022-12-02)
+
+**Note:** Version bump only for package @bolt/uikit-workshop
+
+
+
+
+
+# [5.8.0](https://github.com/bolt-design-system/bolt/tree/master/packages/uikit-workshop/compare/v5.7.5...v5.8.0) (2022-10-18)
+
+**Note:** Version bump only for package @bolt/uikit-workshop
+
+
+
+
+
 # [5.5.0](https://github.com/bolt-design-system/bolt/tree/master/packages/uikit-workshop/compare/v5.4.0...v5.5.0) (2022-05-23)
 
 **Note:** Version bump only for package @bolt/uikit-workshop